Check for existing technologies before query
The resolve function for location.organizations checks to see if
technologies have also been requested. If so, to save some HTTP overhead
it scoops up the org and the technologies in a single query.

Adding a resolve function to organization.technologies ended up
clobbering technologies that had been added by that single query.

This change adds a check for existing technologies and only runs the
query if source.technologies is undefined.