Upsert match only on orgname
The purpose of the upsert is to try to prevent duplication. If there is
an existing organization we can add the new data to it.
Using all the data we have been given for the upsert match means that if
ANY of the data is different (maybe this time there is a founding year
included, or the url now includes a path), the match will fail and a new
org will be created.

Matching only on the org name isn't perfect (Shopify vs Shopify Inc vs
Shopify Inc.) but it should prevent a few more duplicates than the
current approach.