Separate type from acceptance criteria
Previously there was a comingling of concerns in the yearType as logic
for assessing if the AST value _was_ a year was right beside logic for
deciding if the given year fell into an acceptable range.

This commit moves the range check into the resolve function, where it
actually belongs.