Use shouldComponentUpdate == false in <Map>
Previously the <Map> component was implmented using React's contexts to
allow child elements to access the map object while having React's
render cascade properly down from parent to child.

Working with React Router it became clear that changes the the map state
will need to be pushed from the <Map> to the URL, triggering a rerender
of the <Map>'s parent and then the <Map> itself. To avoid this
circularity, shouldComponentUpdate now returns false leaving the <Map>
to manage it's own DOM.

Having dropped contexts which allowed child elements to easily access the
parent child elements like <FlashMessage>, <Layer> and <Navigation> have all
been folded back into the <Map> component.

As part of this work it also made sense to replace `createOrganizationView`
with a proper React component which means this commit also closes #4.