Implement new Map api
This commit moves the logic for determining whether it needs data into
the <Map>. The <Map> component now has two functions:
onDataNeeded and onBoundsChanged
onBoundsChanged gets called every time the map bounds change. Exactly
what you want for updating a URL or a summary.

onDataNeeded is called when the current bounds are larger than the
largest previous bounds.
As an example, when the page is first loaded, onDataNeeded would be
called. Zooming in means the current bounds are inside the bounds we
fetched data for, so no new data is needed.
Panning or zooming outside the bounds we fetched data for causes
onDataNeeded to be called again.