Add config for Google Cloud Build service
This commit adds config files for the cloud build service.  Cloud build both
runs the tests and builds/pushes images.  Since the image is empty and
configured specifically with the collections and indexes needed for running my
tests, is seems silly to expend effort in protecting access to it. Therefore
I'm also including the credentials for my arangodb test image.